Frigide Recipe Balance
Made it so it takes EOW and BOC drops so you can speedrun hellstone. Also some minor changes like the git ignore and moving the BootRecipes.
